+++ crates/engine/body-ir/src/resolution/query/type_path.rs
@@ -4,18 +4,15 @@
 //! query checks those scopes first and then falls back to the semantic/def-map context.
 
 use rg_ir_model::{
-    AssocItemId, DefId, DefMapRef, ImplRef, ItemOwner, ModuleId, ModuleRef, Path, PathSegment,
-    ScopeId, SemanticItemRef, TypeAliasRef, TypePathResolution,
-    items::{TypePath, TypeRef},
+    DefId, DefMapRef, ModuleId, ModuleRef, Path, PathSegment, ScopeId, SemanticItemRef,
+    TypePathResolution,
 };
 use rg_ir_storage::{DefMapSource, ItemStoreSource, NameResolutionFilter, TypePathContext};
 use rg_package_store::PackageStoreError;
 use rg_std::UniqueVec;
-use rg_ty::{GenericArg, NominalTy, Ty, TypeSubst};
+use rg_ty::Ty;
 
 use crate::resolution::BodyResolutionContext;
-
-use super::type_ref::TypeRefUseSite;
 
 pub struct BodyTypePathQuery<'query, D, I> {
     context: BodyResolutionContext<'query, D, I>,
@@ -41,7 +38,11 @@ where
                 Ty::from_type_path_resolution(prefix_resolution, Vec::new()).unwrap_or(Ty::Unknown);
             let mut aliases = UniqueVec::new();
             for ty in prefix_ty.as_nominals() {
-                if let Some(alias) = self.associated_type_alias_for_type(ty, name)? {
+                if let Some(alias) = self
+                    .context
+                    .type_aliases()
+                    .associated_alias_for_type(ty, name)?
+                {
                     aliases.push(alias);
                 }
             }
@@ -172,160 +173,6 @@ where
             TypePathResolution::Unknown
         }
     }
-
-    pub(super) fn ty_from_type_aliases(
-        &self,
-        aliases: &[TypeAliasRef],
-        args: &[GenericArg],
-        subst: &TypeSubst,
-    ) -> Result<Ty, PackageStoreError> {
-        if aliases.len() != 1 {
-            return Ok(Ty::Unknown);
-        }
-
-        self.ty_from_type_alias(
-            aliases
-                .first()
-                .copied()
-                .expect("one alias should exist after length check"),
-            args,
-            subst,
-        )
-    }
-
-    fn ty_from_type_alias(
-        &self,
-        alias_ref: TypeAliasRef,
-        args: &[GenericArg],
-        subst: &TypeSubst,
-    ) -> Result<Ty, PackageStoreError> {
-        let item_query = self.context.item_query();
-        let Some(alias_data) = item_query.type_alias_data(alias_ref)? else {
-            return Ok(Ty::Unknown);
-        };
-        let Some(aliased_ty) = alias_data.signature.aliased_ty() else {
-            return Ok(Ty::Unknown);
-        };
-        if aliased_ty.is_self_type() {
-            return Ok(Ty::Unknown);
-        }
-
-        let mut alias_subst = subst.clone();
-        if let Some(generics) = alias_data.signature.generics() {
-            alias_subst.extend(TypeSubst::from_generics(generics, args));
-        }
-
-        let context = item_query
-            .type_path_context_for_owner(alias_ref.origin, alias_data.owner)?
-            .unwrap_or_else(|| TypePathContext::module(self.context.body().owner_module()));
-        self.context
-            .type_refs(TypeRefUseSite::OwnerContext(context))
-            .with_subst(&alias_subst)
-            .resolve(aliased_ty)
-    }
-
-    /// Attempts to find the type alias with given name for the provided type
-    /// in the body-local context.
-    pub(super) fn associated_type_alias_for_type(
-        &self,
-        ty: &NominalTy,
-        name: &str,
-    ) -> Result<Option<TypeAliasRef>, PackageStoreError> {
-        // In order to find the associated type, we need to iterate through its impl
-        // blocks.
-        let impls = self
-            .context
-            .body_local_items()
-            .inherent_impls_for_type(ty.def)?;
-
-        let item_query = self.context.item_query();
-        for impl_ref in impls {
-            let Some(impl_data) = item_query.impl_data(impl_ref)? else {
-                continue;
-            };
-            if !self
-                .context
-                .impl_matcher()
-                .impl_applies_to_receiver(impl_ref, impl_data, ty)?
-            {
-                continue;
-            }
-
-            for item in &impl_data.items {
-                let AssocItemId::TypeAlias(id) = item else {
-                    continue;
-                };
-                let alias_ref = TypeAliasRef {
-                    origin: impl_ref.origin,
-                    id: *id,
-                };
-                let Some(alias_data) = item_query.type_alias_data(alias_ref)? else {
-                    continue;
-                };
-                if alias_data.name == name {
-                    return Ok(Some(alias_ref));
-                }
-            }
-        }
-
-        Ok(None)
-    }
-
-    pub(super) fn ty_from_associated_type_alias(
-        &self,
-        alias_ref: TypeAliasRef,
-        receiver_ty: &NominalTy,
-        args: &[GenericArg],
-    ) -> Result<Ty, PackageStoreError> {
-        let item_query = self.context.item_query();
-        let Some(alias_data) = item_query.type_alias_data(alias_ref)? else {
-            return Ok(Ty::Unknown);
-        };
-        let Some(aliased_ty) = alias_data.signature.aliased_ty() else {
-            return Ok(Ty::Unknown);
-        };
-        if aliased_ty.is_self_type() {
-            return Ok(Ty::nominal([receiver_ty.clone()].into_iter().collect()));
-        }
-
-        let mut alias_subst = self.semantic_type_subst(receiver_ty)?;
-        if let ItemOwner::Impl(impl_id) = alias_data.owner {
-            let impl_ref = ImplRef {
-                origin: alias_ref.origin,
-                id: impl_id,
-            };
-            if let Some(impl_data) = item_query.impl_data(impl_ref)? {
-                alias_subst.extend(
-                    self.context
-                        .impl_matcher()
-                        .impl_self_subst_for_impl(impl_data, receiver_ty),
-                );
-            }
-        }
-        if let Some(generics) = alias_data.signature.generics() {
-            alias_subst.extend(TypeSubst::from_generics(generics, args));
-        }
-
-        let context = item_query
-            .type_path_context_for_owner(alias_ref.origin, alias_data.owner)?
-            .unwrap_or_else(|| TypePathContext::module(self.context.body().owner_module()));
-        self.context
-            .type_refs(TypeRefUseSite::OwnerContext(context))
-            .with_subst(&alias_subst)
-            .resolve(aliased_ty)
-    }
-
-    pub(super) fn semantic_type_subst(
-        &self,
-        ty: &NominalTy,
-    ) -> Result<TypeSubst, PackageStoreError> {
-        Ok(self
-            .context
-            .item_query()
-            .generic_params_for_type_def(ty.def)?
-            .map(|generics| TypeSubst::from_generics(generics, &ty.args))
-            .unwrap_or_else(TypeSubst::new))
-    }
 }
 
 pub(super) fn split_associated_path(path: &Path) -> Option<(Path, &str)> {
@@ -344,17 +191,4 @@ pub(super) fn split_associated_path(path: &Path) -> Option<(Path, &str)> {
         },
         last_segment.as_str(),
     ))
-}
-
-pub(super) fn prefix_type_ref(path: &TypePath) -> Option<TypeRef> {
-    let prefix_len = path.segments.len().checked_sub(1)?;
-    if prefix_len == 0 {
-        return None;
-    }
-
-    Some(TypeRef::Path(TypePath {
-        source_span: path.source_span,
-        absolute: path.absolute,
-        segments: path.segments[..prefix_len].to_vec(),
-    }))
 }
